Saves as csv pas data to `<output-folder>/pas_details_dump.csv`

The header is as follows:
`solID,pas,target,fuzzy`

+++ easier-uml/src/main/java/it/univaq/disim/sealab/metaheuristic/Launcher.java
@@ -410,7 +410,7 @@ public class Launcher {
 
 		List<RProblem<UMLRSolution>> rProblems = new ArrayList<>();
 		
-		float probPas = Configurator.eINSTANCE.getProbPas();
+		double probPas = Configurator.eINSTANCE.getProbPas();
 
 //		for (Integer eval : Configurator.eINSTANCE.getMaxEvaluation()) {
 		for (Integer l : Configurator.eINSTANCE.getLength()) {
